requirements.txt
Django
pillow
autopep8
djangorestframework
drf-spectacular